8.1.Property Caching{Navigating the Call Graph: Top to Bottom, Bottom to Top; Navigating Dimensions}
Now that we've got our application in production mode and we've dumped the autoloader, it's easier to see what the biggest performance problem is on this page: https://bit.ly/sf-bf-profile4
And actually, there might not be any more problems worth solving. I mean, it's loading in 104 milliseconds... even with the Probe doing all the profiling work.
But... let's see for sure. The function with the highest exclusive time now is PDOStatement::execute()... which is a low-level function that executes SQL queries.

 Tip

The SQL Query information requires a Profiler plan or higher.

If we hover over the query info, these are only taking 12.5 milliseconds... but we are making 43 SQL calls on this page. Is that a problem? It's not ideal, but is it worth fixing? I guess it depends on how much you care... and whether the fix would be easy or if it would add a lot of complexity to our app.

Navigating the Call Graph: Top to Bottom, Bottom to Top
When you're trying to identify where the problem is, there are two ways to look at the call graph - and I often do both to help me understand what's going on. First, you can read from top to bottom - trace through your whole application flow to figure out what's going on down the hot path. Or, you can do the opposite: start at the bottom - start where the problem is... and trace up to find where your code starts.
Let's start from the top: handleRaw() is the framework booting up... and as we trace down... it renders our controller, renders our template... and we're once again inside the body block. This is really the same as last time! Our AppExtension::getUserActivityText() calls the countRecentCommentsForUser() function 23 times. That makes sense: we probably have 23 comments on the page... and for each comment, we need to count all the author's comments to print out this label.

Navigating Dimensions
Before we think about if, and how we might fix this, let's back up and look at other dimensions of this profile. In addition to wall time, we can completely re-draw the call graph based on only I/O time or CPU time. Remember, wall time is I/O time + CPU time. Or we could do something totally different: look at which functions are using the most memory... or even the most network bandwidth.
When we look at this in the network dimension, PDOStatement::execute() - the function that makes SQL calls - shows up here as a big problem. That's because SQL queries are technically network requests.
Re-draw the call graph for the I/O Wait time dimension. We see the same problem here because network calls - and so SQL calls - are part of I/O wait time.
The point is: while "wall time" is typically the most useful dimension, don't forget about these other ones: they can give us more information about what's going on. Is a function slow because of inefficient code inside? Or is it, for example, because of a network call?
Click back to I/O wait time - PDOStatement::execute() is definitely the issue according to this - and the critical path is pretty clear. This one function is taking over half the I/O wait time... but that's only 6 milliseconds. Optimizing this might not be worth it... but let's at least see if we can figure out how to call it less times.
As we already discovered, the problem is coming from CommentRepository::countForUser() which is called by AppExtension::getUserActivityText().
Over in src/Twig/AppExtension.php, each time we render a comment, it calls countForUser() and passes the User object attached to this comment:

 46 lines  src/Twig/AppExtension.php
class AppExtension extends AbstractExtension
{
    public function getUserActivityText(User $user): string
    {
        $commentCount = $this->commentHelper->countRecentCommentsForUser($user);
    }
}