8.2.Property Caching{Property Caching}
Property Caching
Can we optimize this? Well... sometimes, the same user will comment many times on the same sighting - like this vborer user. When that happens, we're making a query to count that user's comments for every comment. That's wasteful!
So here's one idea: leverage "property caching". Basically, we'll keep track of the "status" strings for each user and use that to avoid calculating the status more than once for a given user.
Start by moving most of the logic into a private function called calculateUserActivityText(): this will have a User argument and return a string.
Next, add a new property to the top of the file: private $userStatuses = [].
Back in the public function, here's the magic: if not isset($this->userStatuses[$user->getId()]), then set it by saying $this->userStatuses[$user->getId()] = $this->calculateUserActivityText($user). At the bottom of the function, return $this->userStatuses[$user->getId()].
This is one of my favorite performance tricks because it has no downside, except for some extra code. If getUserActivityText() is called and passed the same User multiple times within a single request, we won't duplicate any work.
So... we probably made our site faster, right? Let's find out! Since we're in Symfony's prod environment, just to be safe, let's clear the cache:

php bin/console cache:clear
and warm it up:

php bin/console cache:warmup
Back in the browser, refresh the page and... let's profile! I'll name this one [Recording] show page try property caching. View the call graph: https://bit.ly/sf-bf-profile-prop-caching.
Ok - PDOStatement still looks like a main problem... but I think we're a little faster. You know what? Let's just compare the two profiles. Go back to the dashboard and compare the previous profile to this one. https://bit.ly/sf-bf-compare-prop-caching. I'll close the old profile.
Ok, so it did help - lower time in each dimension... and we saved 5 queries. So, this is a win, right? Maybe. If you profiled other Big foot sighting pages, which I did, you would find that this often did not help... or helped very little. In fact, this is the first time I've seen it help nearly this much.
So, does the improvement justify the added complexity in our code? If we can repeat this 13% improvement consistently, yea, it is. But if it's more like 1%, probably not.
And even 13% is not that much... and PDOStatement::execute() is still the biggest problem. I feel like the profile is trying to ask us: is there a better way to optimize this?
Next, let's try another approach: using a real cache layer. Truly caching things has its own downside: added complexity in your code and possibly - depending on what you're caching - the need to worry about invalidating cache. We'll want to be sure it's worth it.